backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
PyJWT==2.6.0
bcrypt==4.0.1
Werkzeug==2.2.3